made generalized code for all legs

+++ generalized_code/horse.h
@@ -13,17 +13,16 @@ private:
 public:
     Servo servoS;
     Servo servoE;
-    horse(int pintoS,float valS,int pintoE,float valE)
+    horse(int pintoS,int valS,int pintoE,int valE)
     {
         servoS.attach(pintoS);
         servoE.attach(pintoE);
         corrS=valS;
         corrE=valE;
     }
-    void setangle(float angleS,float angleE)
+    void setangle(int angleS,int angleE)
     {
         servoS.writeMicroseconds(angle(angleS+corrS));
         servoE.writeMicroseconds(angle(angleE+corrE));
     }
 };
-
